Tf idf variations

+++ neo4j-ir-procedures/src/main/java/improvedSearch/BM25.java
@@ -1,7 +1,7 @@
 package improvedSearch;
 
+import model.CardKeyword;
 import result.ResultInfo;
-import model.corpus.CardKeyword;
 import model.Document;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Node;
@@ -48,7 +48,7 @@ public class BM25 {
             res.forEachRemaining(n -> result.put((Long)((Node) n).getProperty("ref"), bm25Score(
                     (String[])((Node) n).getProperty("terms"),
                     (double[])((Node) n).getProperty("idf"),
-                    (int[])((Node) n).getProperty("tf"),
+                    (double[])((Node) n).getProperty("tf"),
                     (int)((Node) n).getProperty("dl"),
                     meanDocumentLength,
                     (double) params.getProperty("k1"),
@@ -73,7 +73,7 @@ public class BM25 {
      * @param query - (Document) query document with query terms as keywords
      * @return - (double) returns the summed bm25 score for all the terms in a document (node)
      */
-    public static double bm25Score(String[] docTerms, double[] idf, int[] tf, int dl, double avgDl, double k1, double b, Document query){
+    public static double bm25Score(String[] docTerms, double[] idf, double[] tf, int dl, double avgDl, double k1, double b, Document query){
 
         // Map with term (String) as key and index of term (Integer) as value
         Map<String, Integer> termPosition = new HashMap<>();
@@ -90,7 +90,7 @@ public class BM25 {
         for(CardKeyword kw : query.keywords){
             if(termsStartsWith(docTerms, kw.getStem())){
                 double tempIdf = idf[termPosition.get(CURRENT_TERM)];
-                int tempTf = tf[termPosition.get(CURRENT_TERM)];
+                double tempTf = tf[termPosition.get(CURRENT_TERM)];
                 sum += tempIdf*((tempTf*(k1+1))/(tempTf+k1*(1-b+(b*(dl/avgDl)))));
 
             }
